requirements.txt
altgraph==0.17.4
asttokens==2.4.1
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
cryptography==43.0.3
debugpy==1.8.7
decorator==5.1.1
et_xmlfile==2.0.0
executing==2.1.0
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.3.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.3
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.4
pefile==2023.2.7
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
Pygments==2.18.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PyMuPDF==1.24.13
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
pyzmq==26.2.0
setuptools==75.3.0
six==1.16.0
stack-data==0.6.3
tk==0.1.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
wcwidth==0.2.13
